main/requirements.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.15
django-jsonfield==1.2.0
django-rest-swagger==2.2.0
djangorestframework==3.11.0
idna==2.8
itypes==1.2.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
nltk==3.4.1
numpy==1.16.2
openapi-codec==1.3.2
pandas==0.23.4
python-dateutil==2.8.1
pytz==2020.1
requests==2.21.0
scikit-learn==0.21.1
scipy==1.5.2
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
uritemplate==3.0.1
urllib3==1.24.3
